+++ Patches/PatchLandmineSize.cs
@@ -28,6 +28,9 @@ namespace RandomEnemiesSize.Patches
 
             var scale = Random.Range(RandomEnemiesSize.instance.minSizeMineEntry.Value,
                 RandomEnemiesSize.instance.maxSizeMineEntry.Value);
+            
+            var customMoonEnemy = RandomEnemiesSize.instance.GetCustomMoonEnemySize(__instance.gameObject.name, StartOfRound.Instance.currentLevel.PlanetName);
+            if (customMoonEnemy.found) scale = Random.Range(customMoonEnemy.minValue, customMoonEnemy.maxValue);
 
             var networkObject = __instance.NetworkObject;